+++ src/file_cache.rs
@@ -64,29 +64,15 @@ enum PinState {
 }
 
 #[derive(Debug)]
-struct PathEntry<T> {
+struct PathEntry {
     /// Meta data corresponding to the most recently inserted entry.
     current: Cell<Option<(PinState, FileMeta)>>,
-    /// The map of entries.
-    entries: InsertMap<FileMeta, Entry<T>>,
 }
 
-impl<T> PathEntry<T> {
-    fn get_or_try_insert<F>(&self, meta: (PinState, FileMeta), init: F) -> Result<&Entry<T>>
-    where
-        F: FnOnce() -> Result<Entry<T>>,
-    {
-        let entry = self.entries.get_or_try_insert(meta.1, init)?;
-        let () = self.current.set(Some(meta));
-        Ok(entry)
-    }
-}
-
-impl<T> Default for PathEntry<T> {
+impl Default for PathEntry {
     fn default() -> Self {
         Self {
             current: Cell::new(None),
-            entries: InsertMap::new(),
         }
     }
 }
@@ -97,10 +83,8 @@ impl<T> Default for PathEntry<T> {
 /// By default all features are enabled.
 #[derive(Clone, Debug)]
 pub(crate) struct Builder<T> {
-    /// Whether to attempt to gather source code location information.
-    ///
-    /// This setting implies usage of debug symbols and forces the corresponding
-    /// flag to `true`.
+    /// Whether or not to automatically reload files that were updated
+    /// since the last open.
     auto_reload: bool,
     /// Phantom data for our otherwise "unused" generic argument.
     _phantom: PhantomData<T>,
@@ -122,6 +106,7 @@ impl<T> Builder<T> {
 
         FileCache {
             cache: InsertMap::new(),
+            entries: InsertMap::new(),
             auto_reload,
         }
     }
@@ -145,9 +130,10 @@ impl<T> Default for Builder<T> {
 /// Note that stale/old entries are never evicted.
 #[derive(Debug)]
 pub(crate) struct FileCache<T> {
-    /// The map we use for associating file meta data with user-defined
-    /// data.
-    cache: InsertMap<PathBuf, PathEntry<T>>,
+    /// The map we use for associating a path with file meta data.
+    cache: InsertMap<PathBuf, PathEntry>,
+    /// The map of entries.
+    entries: InsertMap<FileMeta, Entry<T>>,
     /// Whether or not to automatically reload files that were updated
     /// since the last open.
     auto_reload: bool,
@@ -160,6 +146,25 @@ impl<T> FileCache<T> {
         Builder::<T>::default()
     }
 
+    fn get_or_insert(&self, path: &Path, path_entry: &PathEntry) -> Result<&Entry<T>> {
+        let stat = stat(path).with_context(|| format!("failed to stat `{}`", path.display()))?;
+        let meta = (PinState::Unpinned, FileMeta::from(&stat));
+
+        let entry = self.entries.get_or_try_insert(meta.1, || {
+            // We may end up associating this file with a potentially
+            // outdated `stat` (which could have changed), but the only
+            // consequence is that we'd create a new entry again in the
+            // future. On the bright side, we save one `stat` call.
+            let file = File::open(path)
+                .with_context(|| format!("failed to open file `{}`", path.display()))?;
+            let entry = Entry::new(file);
+            Ok(entry)
+        })?;
+        let () = path_entry.current.set(Some(meta));
+
+        Ok(entry)
+    }
+
     /// Retrieve the entry for the file at the given `path`.
     pub(crate) fn entry(&self, path: &Path) -> Result<(&File, &OnceCell<T>)> {
         let path_entry = self
@@ -169,25 +174,13 @@ impl<T> FileCache<T> {
             if !self.auto_reload || pin_state == PinState::Pinned {
                 // SANITY: Our invariant states that if there is a
                 //         `PathEntry::current` a corresponding entry
-                //         must be in `PathEntry::entries`.
-                let current = path_entry.entries.get(&current_meta).unwrap();
+                //         must be in `FileCache::entries`.
+                let current = self.entries.get(&current_meta).unwrap();
                 return Ok((&current.file, &current.value))
             }
         }
 
-        let stat = stat(path).with_context(|| format!("failed to stat {}", path.display()))?;
-        let meta = (PinState::Unpinned, FileMeta::from(&stat));
-        let entry = path_entry.get_or_try_insert(meta, || {
-            // We may end up associating this file with a potentially
-            // outdated `stat` (which could have changed), but the only
-            // consequence is that we'd create a new entry again in the
-            // future. On the bright side, we save one `stat` call.
-            let file = File::open(path)
-                .with_context(|| format!("failed to open file {}", path.display()))?;
-            let entry = Entry::new(file);
-            Ok(entry)
-        })?;
-
+        let entry = self.get_or_insert(path, path_entry)?;
         Ok((&entry.file, &entry.value))
     }
 
@@ -224,10 +217,16 @@ impl<T> Default for FileCache<T> {
 mod tests {
     use super::*;
 
+    #[cfg(linux)]
+    use std::fs::remove_file;
+    use std::fs::write;
     #[cfg(feature = "nightly")]
     use std::hint::black_box;
     use std::io::Read as _;
-    use std::io::Write as _;
+    #[cfg(linux)]
+    use std::os::fd::AsRawFd as _;
+    #[cfg(linux)]
+    use std::os::unix::fs::symlink;
     use std::thread::sleep;
     use std::time::Duration;
 
@@ -271,14 +270,11 @@ mod tests {
         }
     }
 
-    /// Check that our `FileCache` does not represent symbolic links
-    /// pointing to the same file as equal entries.
+    /// Check that our `FileCache` deduplicates symbolic link targets
+    /// properly.
     #[cfg(linux)]
     #[test]
-    fn file_symlinks() {
-        use std::os::fd::AsRawFd as _;
-        use std::os::unix::fs::symlink;
-
+    fn symlink_entries() {
         let tmpfile = NamedTempFile::new().unwrap();
         let tmpdir = tempdir().unwrap();
         let link = tmpdir.path().join("symlink");
@@ -288,10 +284,101 @@ mod tests {
         let (file1, cell) = cache.entry(tmpfile.path()).unwrap();
         let () = cell.set(42).unwrap();
 
+        // The entry for the link should reference the targeted file.
         let (file2, cell) = cache.entry(&link).unwrap();
-        assert_eq!(cell.get(), None);
+        assert_eq!(cell.get(), Some(&42));
+        assert_eq!(file2.as_raw_fd(), file1.as_raw_fd());
 
-        assert_ne!(file1.as_raw_fd(), file2.as_raw_fd());
+        // Now replace the link with a proper file. This file should
+        // subsequently get picked up.
+        let () = remove_file(&link).unwrap();
+        let () = write(&link, b"test").unwrap();
+        let (file3, cell) = cache.entry(&link).unwrap();
+        assert_eq!(cell.get(), None);
+        assert_ne!(file3.as_raw_fd(), file1.as_raw_fd());
+
+        // But of course the original entry should still exist.
+        let (file4, cell) = cache.entry(tmpfile.path()).unwrap();
+        assert_eq!(cell.get(), Some(&42));
+        assert_eq!(file4.as_raw_fd(), file1.as_raw_fd());
+    }
+
+    /// Make sure that symbolic link chain updates are picked up.
+    #[cfg(linux)]
+    #[test]
+    fn multi_symlink_reload() {
+        // We create the following symbolic link setup:
+        //   link1 -> link2 -> file
+        let tmpfile = NamedTempFile::new().unwrap();
+        let tmpdir = tempdir().unwrap();
+        let link2 = tmpdir.path().join("symlink2");
+        let () = symlink(tmpfile.path(), &link2).unwrap();
+        let link1 = tmpdir.path().join("symlink1");
+        let () = symlink(&link2, &link1).unwrap();
+
+        let cache = FileCache::<usize>::default();
+        let (file1, cell) = cache.entry(&link1).unwrap();
+        let () = cell.set(41).unwrap();
+
+        // Now replace `link2` with a link to a different file.
+        let tmpfile2 = NamedTempFile::new().unwrap();
+        let () = remove_file(&link2).unwrap();
+        let () = symlink(tmpfile2.path(), &link2).unwrap();
+
+        // Our `FileCache` should pick up the change and create a new
+        // entry.
+        let (file2, cell) = cache.entry(&link1).unwrap();
+        assert_eq!(cell.get(), None);
+        assert_ne!(file2.as_raw_fd(), file1.as_raw_fd());
+    }
+
+    /// Check pinning works correctly in the presence of symbolic links.
+    #[cfg(linux)]
+    #[test]
+    fn symlink_pinning() {
+        let tmpfile = NamedTempFile::new().unwrap();
+        let tmpdir = tempdir().unwrap();
+        let link = tmpdir.path().join("symlink");
+        let () = symlink(tmpfile.path(), &link).unwrap();
+
+        let cache = FileCache::<usize>::default();
+        let (file1, cell) = cache.entry(&link).unwrap();
+        let () = cell.set(42).unwrap();
+
+        let () = cache.pin(&link).unwrap();
+
+        // Update symbolic link to point to new file.
+        let tmpfile2 = NamedTempFile::new().unwrap();
+        let () = remove_file(&link).unwrap();
+        let () = symlink(tmpfile2.path(), &link).unwrap();
+
+        // We should still see the pinned content.
+        let (file2, cell) = cache.entry(&link).unwrap();
+        assert_eq!(cell.get(), Some(&42));
+        assert_eq!(file2.as_raw_fd(), file1.as_raw_fd());
+
+        // Update the target file as well and check again.
+        let () = write(tmpfile.path(), b"new-content").unwrap();
+        let (file3, cell) = cache.entry(&link).unwrap();
+        assert_eq!(cell.get(), Some(&42));
+        assert_eq!(file3.as_raw_fd(), file1.as_raw_fd());
+    }
+
+    /// Check that the `FileCache` reports the expected error when
+    /// encountering a symbolic link that points to a non-existent
+    /// target.
+    #[cfg(linux)]
+    #[test]
+    fn symlink_dead_target() {
+        let tmpfile = NamedTempFile::new().unwrap();
+        let tmpdir = tempdir().unwrap();
+        let link = tmpdir.path().join("symlink");
+        let () = symlink(tmpfile.path(), &link).unwrap();
+        let () = tmpfile.close().unwrap();
+
+        let cache = FileCache::<usize>::default();
+        let err = cache.entry(&link).unwrap_err();
+        assert_eq!(err.kind(), ErrorKind::NotFound);
     }
 
     /// Make sure that a changed file updates the cache entry.
@@ -321,8 +408,7 @@ mod tests {
             let () = drop(tmpfile);
 
             {
-                let mut _file = File::create_new(&path).unwrap();
-                let () = _file.write_all(b"foobar").unwrap();
+                let () = write(&path, b"foobar").unwrap();
             }
 
             {